google_search/requirements.txt
google-api-python-client>=2.0.0
requests>=2.25.0
pyyaml>=5.4.0